superhero + citizen

+++ spring/batman/src/main/java/sda/spring/service/SuperHeroService.java
@@ -2,16 +2,21 @@ package sda.spring.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import sda.spring.model.BatmanException;
+import sda.spring.model.Citizen;
 import sda.spring.model.SuperHero;
+import sda.spring.repository.CitizenRepository;
 import sda.spring.repository.SuperHeroRepository;
 
 @Service
 public class SuperHeroService {
     private SuperHeroRepository superHeroRepository;
+    private CitizenRepository citizenRepository;
 
     @Autowired
-    public SuperHeroService(SuperHeroRepository superHeroRepository) {
+    public SuperHeroService(SuperHeroRepository superHeroRepository, CitizenRepository citizenRepository) {
         this.superHeroRepository = superHeroRepository;
+        this.citizenRepository = citizenRepository;
     }
 
     public SuperHero findById(Long id) {
@@ -32,6 +37,23 @@ public class SuperHeroService {
     }
 
     public void deleteById(Long id) {
+        SuperHero hero = superHeroRepository.findOne(id);
+        if (hero != null && hero.getName().equals("Batman")){
+            throw new BatmanException("You shall not remove Batman!");
+        }
         superHeroRepository.delete(id);
+    }
+
+    public void link(String superHeroName, String citizenName){
+        SuperHero superHero = superHeroRepository.findByName(superHeroName);
+        Citizen citizen = citizenRepository.findByName(citizenName);
+        if (citizen == null || superHero == null){
+            throw new IllegalArgumentException("citizen or hero not found");
+        }
+        superHero.getCitizens().add(citizen);
+        citizen.getHeroes().add(superHero);
+        superHeroRepository.save(superHero);
+        citizenRepository.save(citizen);
+
     }
 }
